a2a/agents/langgraph_agent/pyproject.toml
[project]
name = "langgraph-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "click>=8.2.1",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-tavily>=0.2.10",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
